+++ Models/Catalogs/DetalleSalidaCombustible.cs
@@ -8,7 +8,7 @@ namespace Models.Catalogs
 {
     public class DetalleSalidaCombustible
     {
-        public int litros_surtidos { get; set; }
+        public float litros_surtidos { get; set; }
         public SalidaCombustible salida_combustible { get; set; }
         public Tanque tanque { get; set; }
         public Pipa pipa { get; set; }
