+++ Models/Catalogs/Tanque.cs
@@ -9,8 +9,8 @@ namespace Models.Catalogs
     {
         public int id { get; set; }
         public string nombre { get; set; }
-        public int capacidad { get; set; }
-        public int litros { get; set; }
+        public float capacidad { get; set; }
+        public float litros { get; set; }
         public Pipa pipa { get; set; }
         public Combustible combustible { get; set; }
         public DateTime timestamp { get; set; }
