+++ Models/VOs/DetalleAbastecimientoPipaVo.cs
@@ -9,7 +9,7 @@ namespace Models.VOs
     {
         public int id { get; set; }
         public string foto_recibo { get; set; }
-        public int litros { get; set; }
+        public float litros { get; set; }
         public int tanque_id { get; set; }
         public int pipa_id { get; set; }
         public int abastecimiento_id { get; set; }
