+++ Models/VOs/TanqueVo.cs
@@ -9,8 +9,8 @@ namespace Models.VOs
     {
         public int id { get; set; }
         public string nombre { get; set; }
-        public int capacidad { get; set; }
-        public int litros { get; set; }
+        public float capacidad { get; set; }
+        public float litros { get; set; }
         public int pipa_id { get; set; }
         public int combustible_id { get; set; }
     }
